Rename default component filename to index.tsx
This reduces duplicate filepath components and with it the overall
import filepath length.
