+++ webui/src/components/CloseBugButton/index.tsx
