+++ webui/src/components/CloseBugWithCommentButton/index.tsx
